+++ src/oph_core.h
@@ -30,7 +30,7 @@
 typedef enum oph_hier { INVALID_HIER, OPH_ALL, OPH_WEEK, OPH_MONTH, OPH_YEAR } oph_hier;
 typedef enum oph_type { INVALID_TYPE, OPH_INT, OPH_LONG, OPH_SHORT, OPH_BYTE, OPH_FLOAT, OPH_DOUBLE, OPH_COMPLEX_INT, OPH_COMPLEX_LONG, OPH_COMPLEX_FLOAT, OPH_COMPLEX_DOUBLE } oph_type;
 typedef enum oph_oper { INVALID_OPER, OPH_COUNT, OPH_MAX, OPH_MIN, OPH_SUM, OPH_AVG, OPH_STD, OPH_VAR, OPH_CMOMENT, OPH_ACMOMENT, OPH_RMOMENT, OPH_ARMOMENT, OPH_QUANTILE, OPH_ARG_MAX,
-	OPH_ARG_MIN, OPH_MAX_ABS, OPH_MAX_MIN, OPH_AVG_REL
+	OPH_ARG_MIN, OPH_MAX_ABS, OPH_MIN_ABS, OPH_MAX_MIN, OPH_AVG_REL, OPH_MEDIAN
 } oph_oper;
 typedef enum oph_ma_oper { INVALID_MA_OPER, OPH_SMA, OPH_EWMA } oph_ma_oper;
 
@@ -221,9 +221,6 @@ int core_oph_count_multi(oph_multistring * byte_array, oph_multistring * result)
 int core_oph_max(oph_stringPtr byte_array, char *result);
 int core_oph_max_multi(oph_multistring * byte_array, oph_multistring * result);
 
-// Return the maximum of absolute values in an array (as result string)
-int core_oph_max_abs_multi(oph_multistring * byte_array, oph_multistring * result);
-
 // Return the minimum value in an array (as result string)
 int core_oph_min(oph_stringPtr byte_array, char *result);
 int core_oph_min_multi(oph_multistring * byte_array, oph_multistring * result);
@@ -235,6 +232,14 @@ int core_oph_arg_max_multi(oph_multistring * byte_array, oph_multistring * resul
 // Return the index of the minimum value in an array (as result string)
 int core_oph_arg_min(oph_stringPtr byte_array, char *result);
 int core_oph_arg_min_multi(oph_multistring * byte_array, oph_multistring * result);
+
+// Return the maximum of absolute values in an array (as result string)
+int core_oph_max_abs(oph_stringPtr byte_array, char *result);
+int core_oph_max_abs_multi(oph_multistring * byte_array, oph_multistring * result);
+
+// Return the minimum of absolute values in an array (as result string)
+int core_oph_min_abs(oph_stringPtr byte_array, char *result);
+int core_oph_min_abs_multi(oph_multistring * byte_array, oph_multistring * result);
 
 // Return the sum of the array values (as result string)
 int core_oph_sum(oph_stringPtr byte_array, char *result);
@@ -307,6 +312,12 @@ int core_oph_arg_max_array(oph_stringPtr byte_arraya, oph_stringPtr byte_arrayb,
 
 // Evaluates the indexes (1 or 2) of the minimum values comparing each elements of two arrays. Return an array containing the indexes
 int core_oph_arg_min_array(oph_stringPtr byte_arraya, oph_stringPtr byte_arrayb, char *result);
+
+// Evaluates the maximum of the absolute values comparing each elements of two arrays. Return an array containing the maximum values
+int core_oph_max_abs_array(oph_stringPtr byte_arraya, oph_stringPtr byte_arrayb, char *result);
+
+// Evaluates the minimum of the absolute values comparing each elements of two arrays. Return an array containing the minimum values
+int core_oph_min_abs_array(oph_stringPtr byte_arraya, oph_stringPtr byte_arrayb, char *result);
 
 // Return the number of elements in byte_array which have distance from value <= distance
 int core_oph_find(oph_stringPtr byte_array, double value, double distance, long *count);