+++ HISTORY.md
@@ -7,6 +7,8 @@
 
 ### Added:
 
+- Operations 'oph_max_abs' and 'oph_min_abs' to oph_aggregate_operator
+- Operations 'oph_max_abs', 'oph_min_abs' and 'oph_median' to oph_reduce, oph_reduce2 and oph_reduce3
 - New primitve oph_matheval
 
 ### Changed